Double precision Orbit and time
The periapsisRadius, eccentricity, and timeOfPeriapsisPassage fields in the Orbit class were updated to use double precision floats instead of single precision. This led to several other changes throughout the codebase.
Other places in the codebase where variables represented periods of time were also updated to use type double. I did a pretty comprehensive search for them all but still might've missed some, but hopefully not.
All the unit tests (apart from the Lambert Solver, which is the entire point of this branch) are still passing.

+++ Assets/Scripts/Orbital Stuff/GravitationalBody.cs
@@ -27,12 +27,12 @@ public class GravitationalBody
     {
         get
         {
-            float periapsisRadius = Mathf.Ceil((Radius + AtmosphereHeight) * 1.05f / 25000f) * 25000f; // The Ceil() call and factors of 25,000 make the default radius round up to a multiple of 25km
-            float eccentricity = 0.2f;
+            double periapsisRadius = Mathf.Ceil((Radius + AtmosphereHeight) * 1.05f / 25000f) * 25000f; // The Ceil() call and factors of 25,000 make the default radius round up to a multiple of 25km
+            double eccentricity = 0.2f;
             Angle inclination = 0f;
             Angle argumentOfPeriapsis = 0f;
             Angle longitudeOfAscendingNode = 0f;
-            float timeOfPeriapsisPassage = 0f;
+            double timeOfPeriapsisPassage = 0f;
 
             return new Orbit(periapsisRadius, eccentricity, inclination, argumentOfPeriapsis, longitudeOfAscendingNode, timeOfPeriapsisPassage, this);
         }
